requirements.txt
fastapi
uvicorn
python-dotenv
langchain-core
langchain-community
langchain-google-genai
pinecone-client
sentence-transformers
pypdf
langgraph
langchain_pinecone